Squid-py service agreements -- sync with develop, fix conflicts, and fix unittests after merge.
