+++ squid_py/aquariuswrapper.py
@@ -37,7 +37,8 @@ class AquariusWrapper(object):
     def list_assets_metadata(self):
         return json.loads(requests.get(self._base_url + '/ddo').content)
 
-    def publish_asset_metadata(self, asset_did, asset_ddo):
+    def publish_asset_metadata(self, asset_ddo):
+        asset_did = asset_ddo.did
         response = requests.post(self._base_url + '/ddo', data=asset_ddo.as_text(), headers=self._headers)
         if response.status_code == 500:
             raise ValueError("This Asset ID already exists! \n\tHTTP Error message: \n\t\t{}".format(response.text))
