+++ squid_py/ocean/asset.py
@@ -3,18 +3,15 @@ import json
 import logging
 
 from squid_py.ddo import DDO
+from squid_py.did import did_to_id, id_to_did
 from squid_py.ocean.ocean_base import OceanBase
-from squid_py.did import (
-    did_to_id,
-    id_to_did,
-)
+from squid_py.services import ServiceTypes
 
-DDO_SERVICE_METADATA_NAME = 'Metadata'
 DDO_SERVICE_METADATA_KEY = 'metadata'
 
 
-class Asset:
-    def __init__(self, asset_id=None, publisher_id=None, price=None, ddo=None):
+class Asset(OceanBase):
+    def __init__(self, ddo, publisher_id=None):
         """
         Represent an asset in the MetaData store
 
@@ -23,28 +20,32 @@ class Asset:
                 - Use this method to manually build an asset
             2. From a json DDO file Asset.from_ddo_json_file()
                 - Create an asset based on a DDO file
-            3. From a dict object.
-                - Create an asset based in a dict file.
+            3.
 
-        :param asset_id: AKA the DID. This is generated by the market contract, on chain!
+        :param did: `did:op:0x...`, the 0x id portion must be of length 64 (or 32 bytes)
         :param publisher_id:
-        :param price:
         :param ddo: DDO instance
-
-        TODO: remove these init variables ? do we need ...
-
-        asset_id - decided on the DID/DDO and hash of the metadata
-        publisher_id - this should be set when publishing
-        price - set when writing to brizo?
-
         """
 
-        self.asset_id = asset_id
-        self.publisher_id = publisher_id
-        self.price = price
         self._ddo = ddo
-        if self._ddo and self._ddo.is_valid:
-            self.asset_id = did_to_id(self._ddo.did)
+        self.publisher_id = publisher_id
+        self.asset_id = None
+        if self._ddo and self._ddo.is_did_assigend() and self._ddo.is_valid:
+            self.assign_did_from_ddo()
+        else:
+            self.asset_id = hashlib.sha256('assetId'.encode('utf-8')).hexdigest()
+
+        OceanBase.__init__(self, self.asset_id)
+
+    def __str__(self):
+        return "Asset {}, publisher: {}".format(self.did, self.publisher_id)
+
+    def assign_did_from_ddo(self):
+        """
+        #TODO: This is a temporary hack, need to clearly define how DID is assigned!
+        :return:
+        """
+        self.asset_id = did_to_id(self._ddo.did)
 
     @property
     def did(self):
@@ -63,9 +64,8 @@ class Asset:
 
     @classmethod
     def from_ddo_json_file(cls, json_file_path):
-        """return a new Asset object from a DDO JSON file"""
-        this_asset = cls(ddo=DDO(json_filename=json_file_path))
-        logging.debug("Asset {} created from ddo file {} ".format(this_asset.asset_id, json_file_path))
+        this_asset = cls(DDO(json_filename=json_file_path))
+        logging.debug("Asset {} created from ddo file {} ".format(this_asset.did, json_file_path))
         return this_asset
 
     @classmethod
@@ -96,7 +96,7 @@ class Asset:
         # add a signature
         private_password = new_ddo.add_signature()
         # add the service endpoint with the meta data
-        new_ddo.add_service(DDO_SERVICE_METADATA_NAME, service_endpoint, values={DDO_SERVICE_METADATA_KEY: metadata})
+        new_ddo.add_service(ServiceTypes.METADATA, service_endpoint, values={DDO_SERVICE_METADATA_KEY: metadata})
         # add the static proof
         new_ddo.add_proof(0, private_password)
         # create the asset object
@@ -106,7 +106,6 @@ class Asset:
 
     @property
     def metadata(self):
-        """return the metadata for this asset"""
         assert self.has_metadata
         return self._get_metadata()
 
@@ -118,12 +117,22 @@ class Asset:
     def _get_metadata(self):
         """ protected property to read the metadata from the DDO object"""
         result = None
-        metadata_service = self._ddo.get_service(DDO_SERVICE_METADATA_NAME)
+        metadata_service = self._ddo.get_service(ServiceTypes.METADATA)
         if metadata_service:
             values = metadata_service.get_values()
             if DDO_SERVICE_METADATA_KEY in values:
                 result = values[DDO_SERVICE_METADATA_KEY]
         return result
+
+    def get_metadata(self):
+        result = None
+        metadata_service = self._ddo.get_service(ServiceTypes.METADATA)
+        if metadata_service:
+            values = metadata_service.get_values()
+            if 'metadata' in values:
+                result = values['metadata']
+        return result
+
 
     @property
     def is_valid(self):
@@ -166,8 +175,37 @@ class Asset:
 
         return
 
+    def publish_metadata(self):
+        pass
+
+    def update_metadata(self):
+        pass
+
+    def retire_metadata(self):
+        pass
+
     def get_service_agreements(self):
         pass
 
-    def __str__(self):
-        return "Asset {}, price: {}, publisher: {}".format(self.asset_id, self.price, self.publisher_id)
+    def getId(self):
+        return self.asset_id
+
+    def generate_did(self):
+        """
+        During development, the DID can be generated here for convenience.
+        """
+        if not self.ddo:
+            raise AttributeError("No DDO object in {}".format(self))
+        if not self.ddo.is_valid:
+            raise ValueError("Invalid DDO object in {}".format(self))
+
+        metadata = self._get_metadata()
+        if not metadata:
+            raise ValueError("No metedata in {}".format(self))
+
+        if not 'base' in metadata:
+            raise ValueError("Invalid metedata in {}".format(self))
+
+        self.asset_id = hashlib.sha256(json.dumps(metadata['base']).encode('utf-8')).hexdigest()
+        self._ddo = self._ddo.create_new('did:op:%s' % self.asset_id)
+
