+++ tests/resources/ddo/ddo_sample_generated_1.json
@@ -1,0 +1,134 @@
+{
+  "@context": "https://w3id.org/did/v1",
+  "id": "did:op:763e53216438957720c49ea6a4ef995a3ebc39ebdfb3f07b2fab5406b0fabe49",
+  "created": "2018-11-13 15:47:27.916355",
+  "publicKey": [
+    {
+      "id": "did:op:763e53216438957720c49ea6a4ef995a3ebc39ebdfb3f07b2fab5406b0fabe49#keys=1",
+      "type": "RsaSignatureAuthentication2018",
+      "publicKeyPem": "-----BEGIN PUBLIC KEY-----\nMIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQC7aDo+6MOxXZMGVtLYzl4k8p2w\nDgN8g/QvkJoiJUdGqMh3Ijj62CPbcVuFKL50TDLLPZ8Dkc+6VBl8bCawsTjciCVV\nvf36q10//dMH0Py/XGctFGrjfHcgA+BFwmVs2x8pUID6uhDQBNpKIOV2nhYJ+Wo8\nApFiG4b8HBT/J49epwIDAQAB\n-----END PUBLIC KEY-----",
+      "owner": "did:op:763e53216438957720c49ea6a4ef995a3ebc39ebdfb3f07b2fab5406b0fabe49#keys=1"
+    }
+  ],
+  "authentication": [
+    {
+      "type": "RsaVerificationKey2018",
+      "publicKey": "did:op:763e53216438957720c49ea6a4ef995a3ebc39ebdfb3f07b2fab5406b0fabe49#keys=1"
+    }
+  ],
+  "service": [
+    {
+      "id": "did:op:763e53216438957720c49ea6a4ef995a3ebc39ebdfb3f07b2fab5406b0fabe49",
+      "type": "Metadata",
+      "serviceEndpoint": "http://myaquarius.org/api/v1/provider/assets/metadata/{did}",
+      "metadata": {
+        "base": {
+          "name": "UK Weather information 2011",
+          "type": "dataset",
+          "description": "Weather information of UK including temperature and humidity",
+          "size": "3.1gb",
+          "dateCreated": "2012-10-10T17:00:000Z",
+          "author": "Met Office",
+          "license": "CC-BY",
+          "copyrightHolder": "Met Office",
+          "encoding": "UTF-8",
+          "compression": "zip",
+          "contentType": "text/csv",
+          "workExample": "423432fsd,51.509865,-0.118092,2011-01-01T10:55:11+00:00,7.2,68",
+          "contentUrls": [
+            "https://testocnfiles.blob.core.windows.net/testfiles/testzkp.zip"
+          ],
+          "links": [
+            {
+              "name": "Sample of Asset Data",
+              "type": "sample",
+              "url": "https://foo.com/sample.csv"
+            },
+            {
+              "name": "Data Format Definition",
+              "type": "format",
+              "AssetID": "4d517500da0acb0d65a716f61330969334630363ce4a6a9d39691026ac7908ea"
+            }
+          ],
+          "inLanguage": "en",
+          "tags": "weather, uk, 2011, temperature, humidity",
+          "price": 10
+        },
+        "curation": {
+          "rating": 0.93,
+          "numVotes": 123,
+          "schema": "Binary Voting"
+        },
+        "additionalInformation": {
+          "updateFrequency": "yearly",
+          "structuredMarkup": [
+            {
+              "uri": "http://skos.um.es/unescothes/C01194/jsonld",
+              "mediaType": "application/ld+json"
+            },
+            {
+              "uri": "http://skos.um.es/unescothes/C01194/turtle",
+              "mediaType": "text/turtle"
+            }
+          ]
+        }
+      }
+    },
+    {
+      "id": "did:op:763e53216438957720c49ea6a4ef995a3ebc39ebdfb3f07b2fab5406b0fabe49",
+      "type": "OpenIdConnectVersion1.0Service",
+      "serviceEndpoint": "https://openid.example.com/"
+    },
+    {
+      "id": "did:op:763e53216438957720c49ea6a4ef995a3ebc39ebdfb3f07b2fab5406b0fabe49",
+      "type": "CredentialRepositoryService",
+      "serviceEndpoint": "https://repository.example.com/service/8377464"
+    },
+    {
+      "id": "did:op:763e53216438957720c49ea6a4ef995a3ebc39ebdfb3f07b2fab5406b0fabe49",
+      "type": "XdiService",
+      "serviceEndpoint": "https://xdi.example.com/8377464"
+    },
+    {
+      "id": "did:op:763e53216438957720c49ea6a4ef995a3ebc39ebdfb3f07b2fab5406b0fabe49",
+      "type": "HubService",
+      "serviceEndpoint": "https://hub.example.com/.identity/did:op:0123456789abcdef/"
+    },
+    {
+      "id": "did:op:763e53216438957720c49ea6a4ef995a3ebc39ebdfb3f07b2fab5406b0fabe49",
+      "type": "MessagingService",
+      "serviceEndpoint": "https://example.com/messages/8377464"
+    },
+    {
+      "id": "did:op:763e53216438957720c49ea6a4ef995a3ebc39ebdfb3f07b2fab5406b0fabe49",
+      "type": "SocialWebInboxService",
+      "serviceEndpoint": "https://social.example.com/83hfh37dj",
+      "description": "My public social inbox",
+      "spamCost": {
+        "amount": "0.50",
+        "currency": "USD"
+      }
+    },
+    {
+      "id": "did:op:763e53216438957720c49ea6a4ef995a3ebc39ebdfb3f07b2fab5406b0fabe49",
+      "type": "BopsService",
+      "serviceEndpoint": "https://bops.example.com/enterprise/"
+    },
+    {
+      "id": "did:op:763e53216438957720c49ea6a4ef995a3ebc39ebdfb3f07b2fab5406b0fabe49",
+      "type": "Consume",
+      "serviceEndpoint": "http://mybrizo.org/api/v1/brizo/services/consume?pubKey=${pubKey}&serviceId={serviceId}&url={url}"
+    },
+    {
+      "id": "did:op:763e53216438957720c49ea6a4ef995a3ebc39ebdfb3f07b2fab5406b0fabe49",
+      "type": "Compute",
+      "serviceEndpoint": "http://mybrizo.org/api/v1/brizo/services/compute?pubKey=${pubKey}&serviceId={serviceId}&algo={algo}&container={container}"
+    }
+  ],
+  "proof": {
+    "type": "RsaSignatureAuthentication2018",
+    "created": "2018-11-13 15:47:28.003480",
+    "creator": "did:op:763e53216438957720c49ea6a4ef995a3ebc39ebdfb3f07b2fab5406b0fabe49#keys=1",
+    "signatureValue": "uYmurxMoFzuXqpjR3w8sZcMVGuMbo6ZYs4ipHJigJwTinlDh1Y1w3ChMISKD08HnED3zhRqeXb7ssxqTi/vlfre3yp6TG47L0cERanFVq+HP0nbiiVCiDy9DLmVvpyH9WBjXUjW5WS4BlnVrZk59rPUOwFj5ZJ7NmYtwIiK0AgI="
+  }
+}
