+++ tests/resources/ddo/ddo_sample_generated_1_private_key.pem
@@ -1,0 +1,15 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIICXAIBAAKBgQC7aDo+6MOxXZMGVtLYzl4k8p2wDgN8g/QvkJoiJUdGqMh3Ijj6
+2CPbcVuFKL50TDLLPZ8Dkc+6VBl8bCawsTjciCVVvf36q10//dMH0Py/XGctFGrj
+fHcgA+BFwmVs2x8pUID6uhDQBNpKIOV2nhYJ+Wo8ApFiG4b8HBT/J49epwIDAQAB
+AoGAAjJVuHC3AShfyvZlO0hT9H/pgZZFy9r/bOKiLoicU24ysf7IhzGAXhfv4tmC
+6KRkq1raDHXUPFK0XRp8Fkca1lRghrRSvzgB2iARikkGrCPL5xEvpfVva3vJ1iep
+xoeU57BYm6CqKgMsOAxzvGWVsKlkG8tt1sWId80J9mo2fjECQQC80W6rY8Ag4Wtd
+3z/mSznny2cQ/qT34FrrjdG4cciQPDfe8duqYNfJYQNJ7OtgoJooMdoFsA8KM/ET
+dnOzxJ3XAkEA/hZHGh32v6u4b2hcI+rXq5UZIeZkQnoi8xzC914qXdk+dEr6bDya
+Mi0p9KW5LcQ59/SAlGnzPgM6E6PkbhoLsQJBAIPOmzfPUU5iraWA88VaXBjAN7l4
+WUaj+jQW8wCThVOShdn+8b3VDq2IKFiRp5nzYsmOS0jPJyCTM8XHWWtzZE8CQFxC
+0JyeJrvgab8iYNq16owmPpL9qogs+0XVusHB0pZzTOdDVn044DC6D5MDHiONLhNo
+F8xMUIPC/jKlf0uvb3ECQHdhBTLDfRRN4w9A/i/mfusfkbtLHYJ9nLaxi9aJ6ERp
+dBFJh+tREX631+Db/1LU+7R5f7TwogEqYxg1+f9++EA=
+-----END RSA PRIVATE KEY-----
