+++ tests/resources/ddo/ddo_sample_generated_2_private_key.pem
@@ -1,0 +1,15 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIICXQIBAAKBgQDF75wif0VDXidnkUTvQqhjc3qF3Xr6vIVWOvuesZ23wVOX+iUW
+afJmp2EMBkz1B3MD/MyobS8hw7v+MoVNaZyjlEukBqXka0L3DUxKjYdv2f5d9ISs
+Kx2C/xoo+aQySr63OUEEroOJgmhbTMlNBWskBkfbYVWVuI7QODFp3UeInwIDAQAB
+AoGAUpmUaZz89YkUHpcz0jjZ9m6ZYcW0bXBAoGSK3oW2kJlf3hsoYPsjig0tZYw3
+j0VljPNxIJQ3x/pbuE+y60Q4dhrhzo6Zg5ubk9eWCDO2qwI/4WnSwJ/P6AfULlK6
+q1MrEpZ6GOaXEIIMtVUbFybRq+iX8gSeaiZF2mqp/fPfUHkCQQDVBIvPw/hS1G0t
+885CQFLAZ2TqNpeUKAEYn+hUy0SfiPQKvR7RFcrCJEMGrkdQNTQ8PinEQWDVWV6p
+HTP6fhxrAkEA7eADRWsuuOWb78ip4cuwWb9gNt5lcLdpUnevspaxbRva2Wc7iPqL
+b47OgW2p5ER8RNlNwChyNgVe+5woDWcRnQJBAMzyeMnG1I5szzadEd9wH636TH+D
+Usc6zgMxJskeYMHVdGc4mgO0ocW0OcCdwQMMjVtAu9KUzGO4fnjbHf1thE0CQD/w
+hvCR7u0kStYZShC3f5+gIWyeOwYhRt4z7bZDa024vwrRBN+8+N3ztO51W8VP41de
+UdflB04eA3n+b4WqvsECQQCOK3GGSTiL4zS1frVWaHh4XMEZiW69dJZwtpRr2o6i
+LyixOaAD5YpqqbqzKL8owIXfkYedsPiJQk25v/8QPFld
+-----END RSA PRIVATE KEY-----
