+++ tests/service_agreement/test_register_service_agreement.py
@@ -7,7 +7,7 @@ from web3 import Web3, HTTPProvider
 
 from squid_py.config import Config
 from squid_py.keeper.utils import get_fingerprint_by_name, hexstr_to_bytes
-from squid_py.ocean import Ocean
+from squid_py.ocean.ocean import Ocean
 from squid_py.service_agreement.register_service_agreement import (
     execute_pending_service_agreements,
     record_service_agreement,
